PIMS-1451 Fulfill Request > java.lang.NumberFormatException For input string: "c879370936e0d15101374bcb2be20006"
git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@11462 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/controllers/org/pih/warehouse/request/FulfillRequestWorkflowController.groovy
@@ -35,7 +35,7 @@ class FulfillRequestWorkflowController {
 				def command = new FulfillmentCommand();
 				if (params.id) {
 					command = 
-						fulfillmentService.getFulfillment(params.id as int, session.user.id as int)
+						fulfillmentService.getFulfillment(params.id, session.user.id)
 				}
 				else {
 					flash.message = "${warehouse.message(code: 'default.not.found.message', args: [warehouse.message(code: 'request.label', default: 'Request'), params.id])}"
